### 2024-09-04 #7175 @mahato098
try this
https://gofile.io/d/yMHabA
select the desired compiler in the installer
