### 2024-09-04 #7175 @arfon
@edwinb12 would you be willing to review this submission for JOSS?
