### 2024-09-04 #7176 @arfon
@EdwinB12 - would you recommend we publish the submission in JOSS?
