### 2024-09-04 #7176 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
